matlab/requirements_covid_federated.txt
# MATLAB folder: COVID federated learning (PyTorch)
# Last updated: 2026-03-30

torch>=2.0.0,<3.0.0
torchvision>=0.15.0,<1.0.0
numpy>=1.23.0,<3.0.0
Pillow>=9.0.0,<12.0.0
matplotlib>=3.7.0,<4.0.0
scipy>=1.10.0,<2.0.0
